requirements.txt
fastapi
python-multipart
uvicorn
openai
chromadb
PyPDF2
Pillow
nltk
python-dotenv
pydantic
pydantic-settings
supabase
httpx
python-jose
passlib
bcrypt
aiofiles
python-magic
llama-index>=0.9.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.0

# Testing dependencies
pytest>=8.3.4
pytest-asyncio>=0.25.0
pytest-cov>=4.1.0
httpx>=0.27.0

llama-index
llama-index-core
llama-parse
llama-index-readers-file

python-dotenv
pypdf
numpy
scikit-learn
google-generativeai
google-cloud-language